requirements.txt
# --- 핵심 프레임워크 ---
langchain
langchain-community
langchain-core
langchain-text-splitters
langchain-openai
langchain-chroma

# --- 벡터 DB & 검색 ---
chromadb
rank_bm25
flashrank

# --- 데이터 처리 ---
pandas
pypdf
rapidfuzz
olefile
python-dotenv
pyyaml

# --- 웹/앱 인터페이스 ---
streamlit
pysqlite3-binary

# --- 기타 필수 유틸 ---
tqdm
numpy<2.0